+++ module/hotbar-macros.js
@@ -108,16 +108,13 @@ async function createTorgEternityMacro(dropData, slot) {
  */
 async function rollItemMacro(itemName, itemType) {
   const speaker = ChatMessage.implementation.getSpeaker();
-  /** @type {TorgeternityActor} */
-  let actor = game.actors.get(speaker.actor) ?? game.actors.tokens[speaker.token];
+  let actor = ChatMessage.implementation.getSpeakerActor(speaker);
   let item = actor ? actor.items.find(item => item.name === itemName && (!itemType || item.type === itemType)) : null;
   if (!item) {
     // Maybe a UUID was passed in?
     item = fromUuidSync(itemName, { strict: false });
-    if (item)
-      actor = item.parent;
-    else
-      return ui.notifications.warn(_loc('torgeternity.notifications.noItemNamed') + itemName);
+    if (!item) return ui.notifications.warn(_loc('torgeternity.notifications.noItemNamed') + itemName);
+    actor = item.parent;
   }
 
   // Trigger the item roll
@@ -128,14 +125,12 @@ async function rollItemMacro(itemName, itemType) {
     case 'firearm':
     case 'heavyweapon':
     case 'specialability-rollable':
-      actor.rollAttack(item);
-      break;
+      return actor.rollAttack(item);
 
     case 'psionicpower':
     case 'miracle':
     case 'spell':
-      actor.rollPower(item);
-      break;
+      return actor.rollPower(item);
 
     default:
       // this will cause the item to be printed to the chat
@@ -166,7 +161,7 @@ async function rollSkillMacro(skillName, attributeName, isInteractionAttack, DND
 
   let customSkill;
   const speaker = ChatMessage.implementation.getSpeaker();
-  const actor = game.actors.get(speaker.actor) ?? game.actors.tokens[speaker.token];
+  const actor = ChatMessage.implementation.getSpeakerActor(speaker);
   const isAttributeTest = skillName === attributeName;
   const isUnarmed = skillName === 'unarmedCombat';
   let skill = null;